application/backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.3.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.17.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.9
dnspython==2.4.2
ecdsa==0.19.0
email-validator==2.0.0.post2
fastapi==0.101.0
fastapi-jwt-auth==0.5.0
filelock==3.17.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
passlib==1.7.4
pipenv==2024.4.1
platformdirs==4.3.6
pluggy==1.2.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.1.1
pydantic_core==2.4.0
PyJWT==1.7.1
pytest==7.4.0
python-dotenv==1.0.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.1
Random-Word==1.0.11
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.0
SQLAlchemy==1.4.47
starlette==0.27.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.29.2
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
WTForms==3.0.1
